# swagger.yml
openapi: 3.0.1
info:
  title: ThinkLab Manager API
  description: REST API for managing system users and related operations.
  contact:
    name: Development Team
    email: dev@thinklab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "0.1"
...
paths:
...
  /licenses:
    post:
      tags:
      - Software License Management
      summary: Register a new software license contract
      description: Purchases and stores a software contract parameter block into inventory
        database bounds.
      operationId: registerLicense
      parameters:
      - name: X-Executor
        in: header
        description: Identifier of the administrative executive executing the contract
          logging
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareLicenseRequest"
        required: true
      responses:
        "201":
          description: Software license contract successfully cataloged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareLicenseResponse"
        "400":
          description: Invalid payload structural attributes or unrecognized enum
            tokens
  /licenses/{licenseHash}:
    get:
      tags:
      - Software License Management
      summary: Retrieve a distinct license configuration
      description: "Fetches comprehensive details, metadata tracking, status nodes,\
        \ and ongoing allocation lists of a contract."
      operationId: getLicenseByHash
      parameters:
      - name: licenseHash
        in: path
        description: The specific corporate license string token identifier hash
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Software license configuration located and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareLicenseResponse"
        "404":
          description: Contract hash target parameter not found in cluster storage
  /licenses/{licenseHash}/allocations:
    post:
      tags:
      - Software License Management
      summary: Allocate a single software license seat
      description: Deploys and links a license seat entry record to a corporate user
        profile or hardware workstation machine asset.
      operationId: allocateLicenseSeat
      parameters:
      - name: licenseHash
        in: path
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        description: Identifier of the supervisor routing the deployment seat link
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SoftwareLicenseAllocationRequest"
        required: true
      responses:
        "200":
          description: Seat successfully allocated and compliance counters updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoftwareLicenseResponse"
        "400":
          description: Compliance capacity ceiling bounds depletion or expired status
            restriction exception triggered
        "404":
          description: Target contract reference hash not found
...
components:
  schemas:
...
    LicenseAllocation:
      required:
      - allocatedAt
      - allocatedBy
      - assetHash
      - userId
      type: object
      properties:
        assetHash:
          type: string
          description: The target workstation machine hardware asset tracking token.
            Nullable if user-bound.
        userId:
          type: string
          description: The employee identifier utilizing the platform seat access.
            Nullable if machine-bound.
        allocatedAt:
          type: string
          description: The precise timestamp when this seat allocation was recorded.
          format: date-time
        allocatedBy:
          type: string
          description: The identity string of the supervisor or automated agent executing
            the assignment.
      description: Immutable embedded subdocument tracking the atomic assignment of
        a software license seat. Must map explicitly against a hardware asset machine
        target or a corporate user entity identity.
    LicenseStatus:
      type: string
      description: Defines the operational compliance lifecycle states for a software
        license record. Annotated with Serdeable for reflection-free JSON serialization
        bounds.
      enum:
      - ACTIVE
      - EXPIRED
      - SUSPENDED
      - DEPLETED
      x-enum-descriptions:
      - "License is active, valid, and available for hardware or user allocations."
      - The license calendar expiration date has been passed. Structural block on
        new seats allocation.
      - "Administrative lock state (e.g., billing disputes, software audit verification\
        \ pauses)."
      - All purchased seats are fully consumed and allocated across infrastructure
        assets.
    LicenseType:
      type: string
      description: Defines the architectural licensing distribution models for corporate
        software assets. Annotated with Serdeable for native compilation and fast
        JSON serialization.
      enum:
      - PER_SEAT
      - ENTERPRISE
      - SAAS_SUBSCRIPTION
      - VOLUME_MAK
      x-enum-descriptions:
      - Bound to a specific physical hardware machine or single active user profile
        seat.
      - Enterprise-wide contract allowing unrestricted deployment within the entire
        organization bounds.
      - Modern Software-as-a-Service monthly or annual recurring cloud subscription
        vector.
      - Multiple Activation Key format allowing a predefined block lock of sequential
        activations.
...
    SoftwareLicenseAllocationRequest:
      required:
      - assetHash
      - userId
      type: object
      properties:
        assetHash:
          type: string
        userId:
          type: string
      description: Inbound API Request payload representing the allocation of a single
        license seat. Contains optional targets to route the software access link
        to a physical machine or employee.
    SoftwareLicenseRequest:
      required:
      - acquisitionDate
      - expirationDate
      - licenseKey
      - licenseType
      - publisher
      - softwareName
      - totalSeats
      type: object
      properties:
        softwareName:
          minLength: 1
          type: string
          x-size-message: Software name cannot be blank
        publisher:
          minLength: 1
          type: string
          x-size-message: Publisher name cannot be blank
        licenseKey:
          minLength: 1
          type: string
          x-size-message: License activation key serial code cannot be blank
        licenseType:
          minLength: 1
          type: string
          x-size-message: License distribution type model cannot be blank
        totalSeats:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
          x-minimum-message: Total bought seats volume capacity must be greater than
            zero
          x-not-null-message: Total seat block size capacity must be supplied
        acquisitionDate:
          type: string
          format: date-time
          x-not-null-message: Acquisition timeline instance timestamp cannot be null
        expirationDate:
          type: string
          format: date-time
      description: Inbound API Request payload holding structural validation parameters
        for purchasing and cataloging a new corporate software license contract.
    SoftwareLicenseResponse:
      required:
      - acquisitionDate
      - allocatedSeats
      - allocations
      - createdAt
      - expirationDate
      - licenseHash
      - licenseKey
      - licenseType
      - publisher
      - softwareName
      - status
      - totalSeats
      - updatedAt
      type: object
      properties:
        licenseHash:
          type: string
        softwareName:
          type: string
        publisher:
          type: string
        licenseKey:
          type: string
        licenseType:
          $ref: "#/components/schemas/LicenseType"
        status:
          $ref: "#/components/schemas/LicenseStatus"
        totalSeats:
          type: integer
          format: int32
        allocatedSeats:
          type: integer
          format: int32
        acquisitionDate:
          type: string
          format: date-time
        expirationDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/LicenseAllocation"
      description: Outbound unified JSON data snapshot representing a complete corporate
        software license profile.
...